First pass to the bulk actions mode
When you are logged in and you select one or more series using the
table checkboxes, you enter a "bulk actions" mode in which you
can:

* Set the status of the selected series
* Set the reviewer of the selected series
* Archive the series

Note that this mode has no "bundle" option. After a brief
conversation with Ross Burton, it turns out that bundles are
a way of manually creating series. Now that series neatly
bundle themselves, according to Ross there is no need for
"creating bundles of bundles".

Signed-off-by: Belen Barros Pena <belen.barros.pena@linux.intel.com>