Remove the 'to do' page
Looks like the to do page is just a list of patches assigned to
me for review. With the arrival of series, review assignment
happens per series, and not per patch. The series page has an
'Assigned to me for review' filter that shows the series assigned
to the logged in user. So, the 'to page' simply becomes a link to
the 'all series' page with the 'assigned to me for review' filter
applied. So, one page less! \0/

Signed-off-by: Belen Barros Pena <belen.barros.pena@linux.intel.com>