Make the patch URLs easier to find
It was initially suggested to create a separate tab for the patch URLs,
to make them easy to spot, since they are important. However, a whole
new tab seems overkill, considering that often you will have a single
URL per patch (a link to a Bugzilla entry).

Instead of creating a whole new tab, rename the 'metadata' tab to
'URLs and metadata'. Then, make the URLs the first thing in that
tab's content. I hope this works.

Signed-off-by: Belen Barros Pena <belen.barros.pena@linux.intel.com>