Add 'delegate' options for patches
We need to keep the ability to delegate review of specific patches
to another reviewer. For that, we need a new 'delegate to' action,
and a 'delegated to' filter.

It seems strange to me that you can delegate patches that are not
assigned to you for review (although you might be able to do this
only if you are a project administrator: I am not sure).

Signed-off-by: Belen Barros Pena <belen.barros.pena@linux.intel.com>