Add 'tests' column and filter to series table
As per this issue

https://github.com/dlespiau/patchwork/issues/118

For consistency, I have reused the labels we use in the series page,
although the table has now taken a certain Christmas tree feeling that
I don't particularly enjoy. I'd love to use simply a dotted colour,
but that approach is bad form in the accessibility front (using just
colour to convey state causes trouble to colour-blind people and needs
to be tagged properly for screen readers).

So, I'll leave if like a Christmas tree for now. I might come back at a
later stage and revisit the whole thing.

Signed-off-by: Belen Barros Pena <belen.barros.pena@linux.intel.com>