Remove bundle patch page
Initially I considered using a separate page to display a patch included
in a bundle. However this creates a lot of duplication, since the page
would be very similar to the patch details page, which shows the patch in
the context of its series.

Link the list of bundle patches to the patch details page instead, and
include a link to the bundle in it as part of the 'URLs & metadata' tab,
which is now called 'URLs, bundles & metadata'.

This also has the nice side effect of showing the series test results that
apply to the patch.

Signed-off-by: Belen Barros Pena <belen.barros.pena@linux.intel.com>