Changes to the bundle patch list
Change the 'delete' action label to make explicit that this button deletes
the selected patches from the bundle only.

Also, add a 'submitter' column to the patches table and information about
the test results applicable to each patch (i.e. the test results for the
series the patch belongs to).

Signed-off-by: Belen Barros Pena <belen.barros.pena@linux.intel.com>